--- siga-oidc/main/resources/client.properties
@@ -1,38 +1,0 @@
-### OpenID Connect client properties ###
-
-# The default OpenID Connect Provider (OP) properties #
-iss = https://homolog.login.sp.gov.br/sts
-jwks_uri = https://homolog.login.sp.gov.br/sts/.well-known/openid-configuration/jwks
-authz_uri = https://homolog.login.sp.gov.br/sts/connect/authorize
-token_uri = https://homolog.login.sp.gov.br/sts/connect/token
-userinfo_uri = https://homolog.login.sp.gov.br/sts/connect/userinfo
-
-
-# The default OpenID Connect Client (RP) properties #
-client_id = sempapel.documentos
-client_secret = 352f2ef3-db68-43ae-beb0-ee675980e367
-redirect_uri = http://10.2.100.202:8080/oidc-client/callBack
-token_endpoint_auth_method = client_secret_basic
-token_endpoint_auth_signing_alg =
-
-response_type = code
-scope_openid = true
-scope_email = true
-scope_profile = false
-scope_phone = false
-scope_address = false
-scope_offline_access = false 
-prompt = default 
-code_challenge_method = if blank is [disable]
-
-max_age = -1 
-# s TimeUnit.SECONDS
-# m TimeUnit.MINUTES
-# h TimeUnit.HOURS
-# d TimeUnit.DAYS
-
-max_age_time_unit = s
-id_token_hint = null
-login_hint = null
-response_mode = null
-
