--- siga-oidc/src/main/java/br/gov/sp/prodesp/siga/client/JavaTeste01.java
@@ -1,17 +1,0 @@
-package br.gov.sp.prodesp.siga.client;
-
-public class JavaTeste01 {
-
-	public static void main(String[] args) {
-		
-		
-		String valorL = "false";
-		
-		Boolean bb = new Boolean(valorL);
-		
-		
-		System.out.println("Resultado :"  + bb);
-
-	}
-
-}
