+++ siga-oidc/src/main/java/br/gov/sp/prodesp/siga/client/OIDCParameters.java
@@ -5,6 +5,7 @@ import java.util.logging.Logger;
 
 /**
  * OpenID Connect OP, RP and request parameters.
+ * 
  */
 public class OIDCParameters {
