+++ siga/src/main/webapp/WEB-INF/properties/client.properties
@@ -11,7 +11,7 @@ userinfo_uri = https://homolog.login.sp.gov.br/sts/connect/userinfo
 # The default OpenID Connect Client (RP) properties #
 client_id = sempapel.documentos
 client_secret = 352f2ef3-db68-43ae-beb0-ee675980e367
-redirect_uri = http://10.2.100.202:8080/oidc-client/callBack
+redirect_uri = http://10.2.100.202:8080/siga/callBack
 token_endpoint_auth_method = client_secret_basic
 token_endpoint_auth_signing_alg =
 
@@ -22,17 +22,11 @@ scope_profile = false
 scope_phone = false
 scope_address = false
 scope_offline_access = false 
-prompt = default 
-code_challenge_method = if blank is [disable]
+prompt = default
+code_challenge_method = null
 
-max_age = -1 
-# s TimeUnit.SECONDS
-# m TimeUnit.MINUTES
-# h TimeUnit.HOURS
-# d TimeUnit.DAYS
-
+max_age = -1
 max_age_time_unit = s
 id_token_hint = null
 login_hint = null
 response_mode = null
-
